% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegAicha-package.R
\docType{package}
\name{ggsegAicha-package}
\alias{ggsegAicha}
\alias{ggsegAicha-package}
\title{ggsegAicha: AICHA Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

AICHA (Atlas of Intrinsic Connectivity of Homotopic Areas) atlas for the 'ggseg' ecosystem. Provides cortical and subcortical 'ggseg_atlas' objects with 2D polygon geometry and 3D vertex or mesh data, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegAicha}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegAicha/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})
}

}
\keyword{internal}
