% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aicha}
\alias{aicha}
\title{AICHA Atlas (Atlas of Intrinsic Connectivity of Homotopic Areas)}
\usage{
aicha()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the AICHA cortical parcellation with 342 regions.
The original volumetric atlas in MNI space was projected to fsaverage
using the CBIG lab's registration fusion. Contains both 2D polygon
geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for
\code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
aicha()
\dontrun{plot(aicha())}
}
\references{
Joliot M, Jobard G, Naveau M, Delcroix N, Petit L, Zago L,
... & Tzourio-Mazoyer N (2015). AICHA: An atlas of intrinsic
connectivity of homotopic areas. \emph{Journal of Neuroscience Methods},
254, 46-59.
\doi{10.1016/j.jneumeth.2015.07.013}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
